+++ app/Services/ServerAutoOnlineService.php
@@ -51,7 +51,7 @@ class ServerAutoOnlineService
         $shouldShow = !$isGfwBlocked
             && !$isGfwHeld
             && !$isReconnectSuspended
-            && $server->ownAvailableStatus() !== Server::STATUS_OFFLINE;
+            && (int) $server->available_status !== Server::STATUS_OFFLINE;
         $shouldClearGfwAutoHidden = $gfwStatus === ServerGfwCheck::STATUS_NORMAL
             && (bool) $server->gfw_auto_hidden;
         $wasShown = (bool) $server->show;